+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 ---
 ## [3.0.1] - 2022-06-21
 ### Added
-- [pyo3] Add CoverCrypt attributes rotation mechanism
+- [pyo3 + JS/bindgen + FFI] Add CoverCrypt attributes rotation mechanism
 ### Changed
 ### Fixed
+- Fix access policy to partitions
 ### Removed
 ---